requirements.txt

datasets
yfinance
groq
python-dotenv
fastapi
uvicorn
transformers
torch
pandas
